Duality app as a consumer chain (#55)
This PR includes changes to make duality app a consumer chain. To do
this, the following changes are introduced:
1. Since duality would not have its own token and governance will be
done through admin module as agreed some modules are removed:
governance, mint, staking, distribution, genutil, monitoring
*note: monitoring module (initially added by ignite) is removed because
it requires higher version of cosmos SDK than interchain security. If
this module is indeed necessary for duality app, it can be
reintroduced once the interchain repo is upgraded to the higher version
of the SDK
2. Admin module is added. Admin module is used for execution of
governance proposal. Accounts registered as admin are able to submit
those proposals. For duality, that will probably be some multisig
address. This module supports whitelisting which is configurable by app
(when instantiating the admin keeper this list fn is provided).
Currently, for duality app, the whitelisting matches agreed requirements
for consumer chain side governance and is added for the modules which
exists currently in duality app. This list can be extended with the
proposal types from custom modules. You can use
proposals_whitelisting_test.go to check validity of the whitelisted
parameter keys. This version of the admin module still does not support
provider-driven governance, but it is planned to be supported in the
future.
3. App based antehandler is added, so that custom ante handler
decorators can be added at any moment if desired.
4. Export genesis is changed to fit consumer chain needs(validator set
is read from consumer module instead of staking module).
5. Reference to ignite are removed because it requires higher version of
cosmos SDK (similar like monitoring module)
6. Tests for interchain security are integrated into duality
7. Keeper tests are changed to use Fund Account method from its package,
not the one from simApp because the one from simApp uses mint module
under the hood
8. Cli tests are not passing since they are using network package from
cosmos SDK which calls staking module to create validators. Those tests
should be adapted.

Co-authored-by: Nicholas Evans <nicholasevans@Nicholass-MacBook-Pro.local>