Consumer chain, part 2 (#79)
This PR relies on changes implemented in PR "Duality app as a consumer
chain" and adds the following improvements:
- Interchain security is upgraded to tag v0.2.0 which should be used by
consumer chains in the test net. This also upgrades Cosmos SDK version
to v0.45.7, which allowed us to bring back import of Ignite CLI package
that was used previously in the code.
- New CLI command is added which allows Duality binary to be started
locally (one node) without having to start the provider chain and the
relayer. This command is only for testing purposes and should be removed
from binary when it gets built for production.
	
Usage example:
		dualityd init [chain-id]
		dualityd add-consumer-section
		[add as many genesis accounts]
		dualityd start
- Disabled ante handler that stops any other transactions except IBC
ones before channel between provider and consumer is opened. This is
also for testing purposes and should be uncommented when the binary is
built for production.
- Tests in x/dex/client/cli directory that actually start the node can
now be run. A similar approach to the one used to start the node locally
is used to fix these tests. Some tests are still failing, but this is
unrelated to consumer changes.
- Software Upgrade proposal type is whitelisted on the consumer side.