Add Makefile (#296)
I put together a Makefile with some basic commands. This is helpful as a
developer so you don't have to manually run Go commands for everything
and also streamlines the build process. The goal is to utilize this
Makefile in `heighliner` so that we can build Docker images for Duality
that can be fetched with a tag inside of the `ibctest` test cases.


https://github.com/orgs/strangelove-ventures/packages?repo_name=heighliner

`install` will compile the binary which will then be located in whatever
directory `GOBIN` references
`build` will compile the binary into a subdirectory in the root named
`bin`
`clean` will delete the `bin` directory
`test` will run all Go tests with the race detector enabled
`ibctest` will be utilized in a future PR to execute the integration/e2e
tests that utilize `ibctest`