Implement IBC swap middleware  (#106)
# Overview
This PR implements the swap IBC middleware as a standalone middleware
module named `ibcswap`. The middleware is wired up in `app.go` by
creating an IBC middleware stack where the swap middleware wraps the
packet forward middleware which itself wraps the transfer module. Then
we register the route for the transfer module to our new middleware
stack.
 
See:
https://github.com/cosmos/ibc-go/blob/main/docs/ibc/middleware/develop.md
 
 # How it works
 
 The middleware stack is setup to look like this from top to bottom:
```
-- channel.OnRecvPacket
-- swap.OnRecvPacket
-- forward.OnRecvPacket
-- transfer.OnRecvPacket
```	

When ics20 packets come into Duality via IBC, the `OnRecvPacket`
callback in the swap middleware will be executed first. If the packet's
`memo` field contains valid json metadata relevant to the swap
middleware as the root key it will first call down into the underlying
app to invoke the transfer modules `OnRecvPacket` callback, this is how
we get the funds over on the Duality side. It will then attempt a swap
given the swap metadata from the packet's `memo` field.

Example json blob:
```
{
  "swap": {
    "creator": "test-1",
    "receiver": "test-1",
    "tokenA": "token-a",
    "tokenB": "token-b",
    "amountIn": "123.000000000000000000",
    "tokenIn": "token-in",
    "minOut": "456.000000000000000000",
    "non-refundable": false,
    "refund-address": "",
    "next": ""
  }
}
```

A packet can be marked `non-refundable` which indicates that if a swap
fails for some reason, the funds should remain on Duality. If a packet
is marked `non-refundable` AND it has a valid Duality address in
`refund-address` then the funds will be moved to this address upon
failure. Otherwise, the funds will remain in the receiver account on
Duality from the original `FungibleTokenPacketData`.

If the swap is successful we check the `next` field from the packet
metadata and if the value is not empty we mutate the
`FungibleTokenPacketData` to now reference the token denom and amount
that was received from the swap. We also reset the `memo` field to be
the value from the `next` field before calling down into the underlying
app which executes the forward if necessary.


# Testing
Integration tests are written utilizing
[interchaintest](https://github.com/strangelove-ventures/interchaintest)
and live in the `interchaintest` directory. To run the tests run `make
ictest` from the root directory of the repo.

Edit: Currently the Duality repo is private which broke our automatic
Docker image builds via heighliner. So in the mean time to test this
feature you will need to ensure you have heighliner installed on your
local machine and then from the root directory of the repo run
`heighliner build -c duality --local` before you attempt to run the
tests via `make ictest`

---------

Co-authored-by: Teddy Knox <teddyknox@gmail.com>
Co-authored-by: Teddy Knox <teddy@duality.xyz>
Co-authored-by: Julian Compagni Portis <jcompagni@gmail.com>
Co-authored-by: Julian Compagni Portis <julian@terrascope.io>